Explicitly load plugins for karma
By default, karma loads plugins from sibling modules that starts with
`karma-*`. When a package manager (yarn in this case) has a version
conflict, node_modules turns into a tree with nested node_modules/
instead of 100% flat. This makes karma not be able to get the module.

This fix makes it explicitly which plugins we want to load, so it can
work in any of the node_modules modes (tree vs flat)