feat: add Flow type checking
Add `flow check` to check for type errors as additional
linting step.

The `.flowconfig` file in the root of AEgir is needed as else
the linting would fail (although AEgir doesn't define any Flow
types). This is intentional as this branch should only be used
with projects using Flow. There this error message is really
helpful.