feat: add grep and colors option to test
Also specify files to run for node tests, as there is no test/node.js
file.

The output for tests on some color setups is unreadable, disabling
colors helps this immensely.

As I started to bring up in
 https://github.com/ipfs/interface-ipfs-core/issues/325, there isn't a
clean way to isolate tests built into the setup currently. Adding the
grep option gives this to us.