feat: prereleases
This extends the prerelease functionality to allow us to specify the prerelease identifier i.e. "rc" in `1.0.0-rc.0` and to allow prereleases for minor and major versions as well as patch versions.

This adds `--preid` parameter to `aegir release` for specifying the prerelease identifier, see:

https://github.com/npm/node-semver/blob/6086e5ae8e8e253b915b3885a809cfa7532d28f7/semver.js#L365-L395

It also allows the following additional `--type`s: "premajor", "preminor" and "prepatch" to allow us to specify what number in the version should be incremented.

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>