feat: add checks for dependencies
You module might list modules as dependencies that are no longer needed
or can be moved from dependencies to dev-dependencies. With this new
`dependency-check` command you will get a list of those modules.

It doesn't always work correctly, e.g. for AEgir, so manually checking
of the results is needed.

Example:

```console
$ npx aegir dependency-check
Checking for missing dependencies of the main library…
Success! All dependencies used in the code are listed in package.json
Checking for unused dependencies of the main library…
Fail! Modules in package.json not used in code: cids, pull-cat, pull-paramap, pull-traverse
Command failed: dependency-check ./package.json --unused --no-dev
Error: Command failed: dependency-check ./package.json --unused --no-dev
    at makeError (/home/vmx/src/pl/aegir/node_modules/execa/index.js:174:9)
    at Promise.all.then.arr (/home/vmx/src/pl/aegir/node_modules/execa/index.js:278:16)
    at process._tickCallback (internal/process/next_tick.js:68:7)
```

On success the output would be like this:

```
Checking for missing dependencies of the main library…
Success! All dependencies used in the code are listed in package.json
Checking for unused dependencies of the main library…
Success! All dependencies in package.json are used in the code
Checking for missing dev-dependencies in tests…
Success! All dependencies used in the code are listed in package.json
Checking for unused dev-dependencies in tests…
Success! All dependencies in package.json are used in the cod
```

Closes #241.