fix: pull in karma-entry.js from current directory
If you are running lerna or another monorepo tool it's common to
hoist your dependencies into a parent directory.

The loading of karma tests assumes aegir is in the node_modules
directory of your project, which may not be the case if it's been
hoisted.

This change loads the `karma-entry.js` file from the current
directory instead of the hardcoded node_modules path to ensure that
you can actually run your browser tests in a monorepo setup.

It also only adds `karam-entry.js` if we haven't specifed custom
files to test.

asdf