feat: require await in functions marked async
Makes any function marked with the `async` keyword that does not include
an `await` fail linting.

This is because the `async` keyword causes the marked function to be wrapped
in a promise which creates a vm microtask which introduces a small amount of
latency.  In hot codepaths this can add up to large amounts of wasted time.

Closes #354