feat: support async hooks
A common use of hooks in aegir is to start and stop an ipfs node for
browser tests to run against.

Since they all use ipfsd-ctl and that is moving to async/await in
ipfs/js-ipfsd-ctl#353 as part of ipfs/js-ipfs#1670, it would be
nice to not have to mix async and callbacks in the hooks.

This PR adds support for async hooks while not breaking existing
callback based ones.