feat: add commands for publishing rc
This change adds two commands:

```console
aegir update-last-successful-build
```

This command updates a branch called `build/last-successful` (configurable)
It's intended to be run on CI after every successful build of master.

It:

1. Checks that it's running on the head (or detached head)  of `master`
1. Creates a temporary branch
1. Creates `yarn.lock` and `npm-shrinkwrap.json`
1. Adds them to git
1. Deletes the remote `build/last-successful` branch if it exists
1. Pushes the new `build/last-successful` to origin

It deletes the remote branch first because merging conflicts is tedious and
this branch should only be used by automated processes.

```console
aegir publish-rc
```

This creates a release branch for the next release and publishes a release
candidate to npm.

It:

1. Deletes your local copy of `build/last-successful`
1. Pulls down and checks out an updated version
1. Creates the release branch by `semver.inc`ing the `package.json` version
1. Overwrites existing pre-id in the version number
1. Uses aegir's release command to publish an `.rc` version