fix: test external modules that have transitive dep on ipfs
Goes some way to address ipfs/js-ipfs#2542 in that it changes our
definition of how a module 'depends' on IPFS.  If it's in your
`node_modules` folder, it'll get upgrade to the current release we
are testing.

This means if we're testing module `A` and `A` has a depdendency
graph like `A -> B -> C -> ipfs`, as long as `ipfs` has been hoisted
to the the top level `node_modules` folder, we'll swap it out with
the new version we're trying to release.

It also dedupes `ipfs` and `ipfs-http-client` in the dep tree so
we can be sure we're testing against our `rc`, though this may
cause problems where intermediate deps need updating, but at least
the need for them to be updated will  be visible.