feat: make unhandled promise rejections fail tests
Adds a new flag to the test runner: `--no-unhandled-promise-rejections`
which defaults to `true`.

When true this causes the `unhandled-rejections` node flag to be set
to `true` during the test run under node.

This means any `unhandledRejection` events raised on `process` will
cause the process to exit and the tests to fail.

This is to ease our transition into the brave new world of mixing
error handling strategies between async and event emitters.

Fixes #505