fix: remove package-lock before generating npm-shrinkwrap file
`npm shrinkwrap` renames your `package-lock.json` if present, this
includes project dev deps which npm seems to be installing even
though the published module with the shrinkwrap file is a dependency.

If we install with the `--production` flag, then remove the
`package-lock.json` file and *then* run shrinkwrap, no dev deps are
included in the shrinkwrap file.

This only affects npm, yarn seems to do the right thing.

Fixes #516