feat: send body back in echo response if sent
If you send a body, echo it back as the `body` property of the response.

Uses `json-stringify-safe` to serialize the request otherwise it doesn't
work because it contains circular references.

Also adds tests for some of the echo server routes as they weren't tested.

Not sure why we need the `/echo/query` route, the basic `/echo` route can
be used instead.