feat: checks action and bundlesize check improvements (#532)
This PR makes the `aegir build --bundlesize` use our own build config and some extra goodies in the output text. `--analyze` was remove in favor of `--bundlesize` doing everything in on cmd.

Included is also a github action to check the bundle size. Check this PR checks below to see how it works.

this new github action can potentially include all the cmds we run in the check job in travis