fix: export types for chai extensions
If a chai extension has types it cannot be resolved unless those types
are in a definitely typed style `@types/chai-something` module.

This seems to be because we don't export them from the `utils/chai.js`
file so they get removed during compilation, then when we try to use
the extensions their types have not been loaded so chai's base types are
not extended and we have missing methods everywhere.

Here we export the chai extensions, ensuring their types will be loaded
then delete the exported property as we don't really want to export the
extensions.