fix: export types for chai extensions (#789)
If a chai extension has types they cannot be resolved by modules that depend
on aegir unless those types are in a definitely-typed-style `@types/chai-foo`
module.

This seems to be because we don't export them from the `utils/chai.js`
file so they are not referenced from anywhere, then when we try to use
chai with the extensions from another module their types have not been
loaded so chai's base types are not extended and we have missing methods
everywhere. Unless of course, they are in a `@types/...` module in which case
typescript's default module resolution algorithm will find them, though more
by accident than by design.

Here we export the chai extensions, ensuring their types will be loaded
by the generated `.d.ts` file, then delete the exported property as we don't
really want to export the extensions.