fix: pin ts version
`typescript@4.4.x` changes the default type for `catch` variables
from `any` to `unknown` which means you have to then use `instanceof`
and other tricks to let `tsc` know what type of error you are handling.

This is a breaking change released as a minor, sigh, so pin `typescript`
to `4.3.x` until we are ready to upgrade.