chore: remove chai-checkmark (#906)
Chai-checkmark doesn't work well with promise based flow control as it requires a `done` function and mocha objects to the use of both `done` and `async` in test function signatures.

Probably better to just use await instead.  As a bonus it also makes the logic in a test easier to follow.