fix: use @semantic-release/github fork
@semantic-release/github doesn't respect GitHub's rate limiting - it
did once but I think GH changed their response type and now it doesn't
which means publishing large modules like `@libp2p/interfaces` is a dice
roll as to whether it will succeed or not.

Here we switch to a temporary fork that uses the `@octokit` plugins
for throttling and retrying so it should be more reliable.

Refs: https://github.com/semantic-release/semantic-release/issues/2204