fix: remove no-bundle from test build and fix flag
We run build before test but pass `--no-bundle` on the assumption
that a project build command is `aegir build`.  Passing `--no-bundle`
skips making a minified build for a tiny speedup.

If the project build command isn't `aegir build` this can break things
so remove the optimisation.

Also build by default before test and fix the problem whereby the build
flag was being ignored for typescript projects.

Fixes: #975