feat: align monorepo sibling versions (#1010)
We use `semantic-release-monorepo` to publish packages in monorepos.

We use `lerna` to work out sibling cross-dependencies to ensure we publish in the correct order.

This all seems to work ok, but `semantic-release-monorepo` doesn't update dependency versions for cross-dependent siblings when they change.

We can lean on semver ranges to take care of this for the most part but it breaks when we release a major version of a sibling.  In this case the other siblings have been coded against the new API but still depend on the old version which breaks.

This pr adds a job to the release command that, if we are a monorepo, synchronises the versions of all sibling dependencies after each publish.