feat: build opt-out for cache-node-modules action
I've used this action several times and did not know that it was also called `npm run build`. 

The build step can increase a workflow's runtime and complexity, and I usually don't want to do that. I think it's useful, but since there was already support for `inputs.build`, I think it's unnecessary. An opt-out (since we don't want to break existing users) will help me out a lot.