feat: link to external typedocs
The TypeDoc docs [describe how to link to third party symbols](https://github.com/TypeStrong/typedoc/blob/master/internal-docs/third-party-symbols.md)
so this PR adds the beginning of support for that.

It adds a plugin with links to the docs of a few well-known types which
will be used to link through to their definitions in generated docs.

It also adds the ability for third party modules to declare their own
docs in their manifests.  For example for `@multiformats/multiaddr` to
make use of the `Multiaddr` interface linkable in the documentation of
other modules, it needs to add the following to it's `package.json`:

```js
{
  // ... other stuff
  "docs": {
    "Multiaddr": "https://multiformats.github.io/js-multiaddr/interfaces/index.Multiaddr.html"
  }
}
```