feat: add release-rc command
The last thing we use lerna for is releasing RC versions of monorepo
packages, so add a command here that does the same thing.

It adds the git hash to the version number, updates sibling dependencies
to that new version number then publishes the packages all in the order
of dependency.