fix: do not enforce engines in package.json
The engines field causes warnings on npm but errors on yarn when the
required node version is not met.

Not all of our modules require LTS node and they are used by projects
outside of libp2p/helia/etc.

This change removes the automatic addition of an engines field to a
project's `package.json` while running the `check-project` command,
instead projects that do require a certain node version are free to
define it and deal with the breakage that subsequently occurs.

Partial revert of: #1184
Refs: https://github.com/ipfs/aegir/issues/1276