fix!: revert to semantic-release-monorepo
`@anolilab/multi-semantic-release` does not cause semantic release
to check the commits of each package before working out if anything
has changed so creates unecessary comments on long-closed github
issues.

`semantic-release-monorepo` did not do this and has been updated to
be ESM-compatible recently so switch back to that.

BREAKING CHANGE: if using `aegir release` in the scripts of a monorepo package, config must be updated:

1. Change `"release": "aegir release"` in the monorepo root to `"release": "aegir run release"`
2. Add `"release": "aegir release"` to each monorepo package
3. Add semantic release config to each monorepo package