fix: ignore module-not-found errors in doc-check
Sometimes we want to show examples that use modules that aren't a
dependency of the module we are documenting.

In this case tsc throws `TS2307` when trying to compile the example
code block.

Ignore this error so we can just treat these deps as `any` while
still getting the benefit of compiling the rest of the code block.